Finished Functions exercises. Began on Structures exercises.

+++ Functions/utils.h
@@ -15,4 +15,4 @@ int clamp(int lower, int upper, int);
 
 float distance(float x1, float y1, float x2, float y2);
 
-int estTimeToArrival(int x1, int y1, int x2, int y2, int speed);
+float estTimeToArrival(float x1, float y1, float x2, float y2, float speed);
